+++ WinQuake/d_fill.c
@@ -48,17 +48,17 @@ void D_FillRect (vrect_t *rect, int color)
 		rheight += ry;
 		ry = 0;
 	}
-	if (rx+rwidth > vid.width)
-		rwidth = vid.width - rx;
-	if (ry+rheight > vid.height)
-		rheight = vid.height - rx;
+	if (rx+rwidth > (int)vid.width)
+		rwidth = (int)vid.width - rx;
+	if (ry+rheight > (int)vid.height)
+		rheight = (int)vid.height - rx;
 		
 	if (rwidth < 1 || rheight < 1)
 		return;
 
 	dest = ((byte *)vid.buffer + ry*vid.rowbytes + rx);
 
-	if (((rwidth & 0x03) == 0) && (((long)dest & 0x03) == 0))
+	if (((rwidth & 0x03) == 0) && (((intptr_t)dest & 0x03) == 0))
 	{
 	// faster aligned dword clear
 		ldest = (unsigned *)dest;
