+++ WinQuake/sys.h
@@ -31,8 +31,8 @@ int Sys_FileOpenRead (char *path, int *hndl);
 int Sys_FileOpenWrite (char *path);
 void Sys_FileClose (int handle);
 void Sys_FileSeek (int handle, int position);
-int Sys_FileRead (int handle, void *dest, int count);
-int Sys_FileWrite (int handle, void *data, int count);
+size_t Sys_FileRead (int handle, void *dest, int count);
+size_t Sys_FileWrite (int handle, void *data, int count);
 int	Sys_FileTime (char *path);
 void Sys_mkdir (char *path);
