+++ WinQuake/zone.h
@@ -86,7 +86,7 @@ Zone block
 void Memory_Init (void *buf, int size);
 
 void Z_Free (void *ptr);
-void *Z_Malloc (int size);			// returns 0 filled memory
+void *Z_Malloc (size_t size);			// returns 0 filled memory
 void *Z_TagMalloc (int size, int tag);
 
 void Z_DumpHeap (void);
